fix: redundant cast in example test (clippy test profile)

+++ crates/pixtuoid/examples/snapshot.rs
@@ -1855,8 +1855,7 @@ mod tests {
         let (scene, warmup_ms) = meeting_scene(now, 3, cols, rows, 0, max_desks, 12).unwrap();
         assert_eq!(scene.agents.len(), 12, "staged 3 + 9 archetype fillers");
 
-        let layout =
-            SceneLayout::compute_with_seed(cols, (rows - 1) as u16 * 2, max_desks, 0).unwrap();
+        let layout = SceneLayout::compute_with_seed(cols, (rows - 1) * 2, max_desks, 0).unwrap();
         let staged: Vec<_> = scene
             .agents
             .values()
